# openapi/openapi-updated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Marin Pest Control Dashboard API",
    "version": "2.0.0",
    "description": "Comprehensive business management system integrating QuickBooks Online, Google Calendar, Supabase authentication, and custom scheduling features for Marin Pest Control",
    "contact": {
      "name": "Marin Pest Control",
      "email": "support@wemakemarin.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://wemakemarin.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "supabaseAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "summary": "Get current authenticated user information",
        "tags": ["Authentication"],
        "security": [{"supabaseAuth": []}],
        "responses": {
          "200": {
            "description": "User information retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout current user",
        "tags": ["Authentication"],
        "security": [{"supabaseAuth": []}],
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "get": {
        "summary": "Refresh user session",
        "tags": ["Authentication"],
        "security": [{"supabaseAuth": []}],
        "responses": {
          "200": {
            "description": "Session refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "session": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "supabaseAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "type": "object" }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "isAdmin": { "type": "boolean" },
          "isActive": { "type": "boolean" },
          "lastLogin": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
}